Move functions into only module that uses them
