+++ packages/compile-common/src/profiler/index.ts
@@ -1,9 +1,4 @@
-export * from "./types";
-
-export { readAndParseArtifactFiles } from "./readAndParseArtifactFiles";
-export { findUpdatedFiles } from "./findUpdatedFiles";
 export { isExplicitlyRelative } from "./isExplicitlyRelative";
-export { minimumUpdatedTimePerSource } from "./minimumUpdatedTimePerSource";
 export { resolveAllSources } from "./resolveAllSources";
 export { requiredSources } from "./requiredSources";
 export { updated } from "./updated";
